Improvements to deleting Entities/Mobys. Previously we could have an issue where we deleted a moby server-side, but the client had not been notified yet and sent us updates relating to the deleted entity. If we'd immediately assigned a new moby, we would think the packet coming in related to that new moby. Now we don't assign the same ID as we just deleted. We should probably send an ack packet to the client and not assign a new Moby to the deleted slot until we have a confirmation from the client.

+++ Lawrence/Player.cs
@@ -137,7 +137,9 @@ namespace Lawrence
 
         public override void OnDeleteEntity(DeleteEntityNotification notification) {
             if (notification.Entity is Moby moby) {
-                _client.DeleteMoby(moby);
+                if (_client.HasMoby(moby)) {
+                    _client.DeleteMoby(moby);
+                }
             }
 
             base.OnDeleteEntity(notification);
